% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolbarplots.R
\name{point_barplot}
\alias{point_barplot}
\title{Barplot made with points}
\usage{
point_barplot(x, y, x.label = "X", y.label = "Y")
}
\arguments{
\item{x}{A factor vector}

\item{y}{A frequency vector}

\item{x.label}{The x axis label}

\item{y.label}{The y axis label}
}
\value{
A ggplot based plot
}
\description{
Plot random points as the bars.
}
